Move dark mode helpers to application helper
