+++ playbook/app/controllers/playbook/application_controller.rb
@@ -2,6 +2,7 @@
 
 module Playbook
   class ApplicationController < ActionController::Base
+    helper Playbook::ApplicationHelper
     helper Playbook::PbDocHelper
     helper Playbook::PbSampleHelper
     append_view_path Playbook::Engine.root + "app/pb_kits"
